# perf(multi_object_tracker): fast IoU calculation in the tracker association and merging