--- a/perception/autoware_multi_object_tracker/lib/object_model/shapes.cpp
+++ b/perception/autoware_multi_object_tracker/lib/object_model/shapes.cpp
@@ -63,16 +63,18 @@ inline double getUnionArea(
 double get1dIoU(
   const types::DynamicObject & source_object, const types::DynamicObject & target_object)
 {
-  static const double min_union_length = 0.1;  // As 0.01 used in 2dIoU, use 0.1 here
-  static const double min_length = 1e-3;       // As 1e-6 used in 2dIoU, use 1e-3 here
+  constexpr double min_union_length = 0.1;  // As 0.01 used in 2dIoU, use 0.1 here
+  constexpr double min_length = 1e-3;       // As 1e-6 used in 2dIoU, use 1e-3 here
   // Compute radii from dimensions (use max of x and y as diameter)
-  double r_src = std::max(source_object.shape.dimensions.x, source_object.shape.dimensions.y) * 0.5;
-  double r_tgt = std::max(target_object.shape.dimensions.x, target_object.shape.dimensions.y) * 0.5;
+  const double r_src =
+    std::max(source_object.shape.dimensions.x, source_object.shape.dimensions.y) * 0.5;
+  const double r_tgt =
+    std::max(target_object.shape.dimensions.x, target_object.shape.dimensions.y) * 0.5;
   // if radius is smaller than the minimum length, return 0.0
   if (r_src < min_length || r_tgt < min_length) return 0.0;
   // Ensure r1 is the larger radius
-  double r1 = std::max(r_tgt, r_src);
-  double r2 = std::min(r_tgt, r_src);
+  const double r1 = std::max(r_tgt, r_src);
+  const double r2 = std::min(r_tgt, r_src);
   const auto dx = source_object.pose.position.x - target_object.pose.position.x;
   const auto dy = source_object.pose.position.y - target_object.pose.position.y;
   // distance between centers
@@ -88,10 +90,9 @@ double get1dIoU(
   // approximation
   const double intersection_length = std::max(0.0, r1 + r2 - dist);
   const double union_length = r1 + r2 + dist;
-  const double iou =
-    union_length < min_union_length
-      ? 0.0
-      : std::min(1.0, intersection_length * intersection_length / (union_length * union_length));
+  const double iou = union_length < min_union_length
+                       ? 0.0
+                       : std::min(1.0, intersection_length * r2 / (r1 * r1) * 0.5);
   return iou;
 }