## Description
- Introduced a distance-based scoring mechanism for pedestrian measurements in the DataAssociation class, bypassing the IoU check.
- Updated the mergeOverlappedTracker method to implement specific merging criteria for overlapping pedestrian trackers, improving accuracy in tracking.
- Added a new helper function, canMergePedestrianTargets, to encapsulate the logic for merging pedestrian targets based on confidence and distance.

These changes enhance the robustness of the multi-object tracking system, particularly for pedestrian detection and association.




### Performance 


#### Test on Local Unit Testing
**before**
=== Overall Performance Statistics ===
Total      - Avg: 4.89509  ms Min: 4.012    ms Max: 16.364   ms

=== Individual Function Performance ===
Predict    - Avg: 0.226358 ms Min: 0        ms Max: 0.808    ms
Associate  - Avg: **3.13088**  ms Min: 0.003    ms Max: 10.558   ms
Update     - Avg: 0.394866 ms Min: 0        ms Max: 1.81     ms
Prune      - Avg: **0.558026** ms Min: 0.015    ms Max: 1.983    ms
Spawn      - Avg: 0.582694 ms Min: 0.279    ms Max: 7.728    ms

**after**
=== Overall Performance Statistics ===
Total      - Avg: 3.62503  ms Min: 2.913    ms Max: 9.944    ms

=== Individual Function Performance ===
Predict    - Avg: 0.202342 ms Min: 0        ms Max: 0.586    ms
Associate  - Avg: **2.22312**  ms Min: 0.002    ms Max: 6.474    ms
Update     - Avg: 0.362214 ms Min: 0        ms Max: 1.401    ms
Prune      - Avg: **0.356458** ms Min: 0.017    ms Max: 0.88     ms
Spawn      - Avg: 0.478646 ms Min: 0.222    ms Max: 8.386    ms

#### Test with Actual Vehicle
The improvement has been tested and confirmed with actual vehicle, with no problems observed during experiment.
- [TIER IV internal link](https://star4.slack.com/archives/C07DMS06H6U/p1751530468044419)


#### Test with Evaluator
- [TIER IV internal link](https://evaluation.tier4.jp/evaluation/reports/eff3285f-3a8f-5bb5-b6ec-ab43f728df5e?project_id=prd_jt)

#### Test with LSIM

Current result
Detect(Yellow) and Track(Green)

https://github.com/user-attachments/assets/a34d8279-d8b7-49be-89b5-c5c0459c64d4



Compared with before
Track Before(Red) and Track After(Green)

https://github.com/user-attachments/assets/15259e2a-9304-4808-b181-39e133a3af8c




## Note 
### Associate part
To clarify two two questions:
1. Should we only consider ped × ped associations?
2. Should we use 1D-IoU instead of 2D-IoU?
To evaluate this, we organized several cases below, comparing three approaches to show the calculations used:
Origin: The current implementation in the main branch (uses 2D-IoU for all cases).
A: The current PR behavior, which bypasses IoU checks for ped × ped and relies only on score.
B: A distance-based IoU alternative, replacing 2D-IoU for pedestrian matching.

Track | Detect | GT IoU | Origin  | A| B | Association Result (Score) Compare
-- | -- | -- | -- | -- | -- | --
ped | ped | 0.5 |2D-IoU×1 Score×1 | Score×1 | 1D-IoU×1 Score×1 | Same
ped | ped | 0 |  2D-IoU×1 | Score×1 |  1D-IoU ×1  | Similar (score ≈ 0 due to large distance)
other | ped | 0.5 | 2D-IoU×1 Score×1 |Score×1 |  2D-IoU×1  Score×1 | Same
other | ped | 0 |  2D-IoU×1 | Score×1 | 2D-IoU×1  | Similar (score ≈ 0 due to large distance)

In the table, expressions like 2D-IoU × 1, Score × 1 indicate that both the IoU and the score estimation are part of the association calculation in the original method.
By contrast, the A method bypasses IoU entirely and uses the score directly, yet the final outcome remains almost the same — because when the IoU is high, the iou calculation doesn’t change the score , and when IoU is 0, the score would have been suppressed anyway.

For pedestrian-to-pedestrian associations:
Skipping the IoU check (as in this PR)produces similar or same results compared to the original approach which has iou calculation .

For non-pedestrian cases:
All methods behave similarly when the ground truth IoU is high. 
When IoU is zero (distant), the resulting scores are close to zero in all methods, leading to no false associations.

Skipping the IoU step for ped associations can improves association performance.


### Merge part

This PR logic only applies to ped × ped cases. (For other × ped, it will fall back to the original flow.)
Currently,  IoU here is for the logic _if  iou<=  target_known? min_known_iou: min_unknown_iou  do not merge_ 
The question is whether we can simplify this for can we simplified ped × ped by replacing the "IoU too small" check with a "distance too far" check?

###



## Related links

**Parent Issue:**

- Link

<!-- ⬇️🟢
**Private Links:**

- [TIER IV internal link](https://star4.slack.com/archives/C07DMS06H6U/p1751530468044419)
⬆️🟢 -->

## How was this PR tested?

## Notes for reviewers

None.

## Interface changes

None.

<!-- ⬇️🔴

### Topic changes

#### Additions and removals

| Change type   | Topic Type      | Topic Name    | Message Type        | Description       |
|:--------------|:----------------|:--------------|:--------------------|:------------------|
| Added/Removed | Pub/Sub/Srv/Cli | `/topic_name` | `std_msgs/String`   | Topic description |

#### Modifications

| Version | Topic Type      | Topic Name        | Message Type        | Description       |
|:--------|:----------------|:------------------|:--------------------|:------------------|
| Old     | Pub/Sub/Srv/Cli | `/old_topic_name` | `sensor_msgs/Image` | Topic description |
| New     | Pub/Sub/Srv/Cli | `/new_topic_name` | `sensor_msgs/Image` | Topic description |

### ROS Parameter Changes

#### Additions and removals

| Change type   | Parameter Name | Type     | Default Value | Description       |
|:--------------|:---------------|:---------|:--------------|:------------------|
| Added/Removed | `param_name`   | `double` | `1.0`         | Param description |

#### Modifications

| Version | Parameter Name   | Type     | Default Value | Description       |
|:--------|:-----------------|:---------|:--------------|:------------------|
| Old     | `old_param_name` | `double` | `1.0`         | Param description |
| New     | `new_param_name` | `double` | `1.0`         | Param description |

🔴⬆️ -->

## Effects on system behavior

None.